QA to Master merge

+++ projects/ui/src/lib/oui/datepicker/native-date-formats.ts
@@ -5,7 +5,7 @@ export const OUI_NATIVE_DATE_FORMATS: OuiDateFormats = {
     dateInput: null
   },
   display: {
-    dateInput: { year: 'numeric', month: 'long', day: 'numeric' },
+    dateInput: { year: 'numeric', month: 'short', day: 'numeric' },
     monthYearLabel: { year: 'numeric', month: 'short' },
     dateA11yLabel: { year: 'numeric', month: 'long', day: 'numeric' },
     monthYearA11yLabel: { year: 'numeric', month: 'long' }
